fix(breadcrumbs): Fix wrapping and alignment
Chains of long breadcrumbs result in a full breadcrumb bar which wraps
on smaller screens. The `max-height` property causes the breadcrumb bar
content to overflow its background, which results in hard-to-see button
text.

This change removes the `max-height` property, allowing the breadcrumb
bar background to grow with a wrapped/multi-line breadcrumb bar.